+++ malachite-float/src/test_util/generators/mod.rs
@@ -6362,6 +6362,17 @@ pub fn rational_unsigned_unsigned_rounding_mode_quadruple_gen_var_8()
     )
 }
 
+// All `(Rational, u64, u64, RoundingMode)` that are valid inputs to
+// `Float::asec_with_period_rational_prec_round`.
+pub fn rational_unsigned_unsigned_rounding_mode_quadruple_gen_var_9()
+-> Generator<(Rational, u64, u64, RoundingMode)> {
+    Generator::new(
+        &exhaustive_rational_unsigned_unsigned_rounding_mode_quadruple_gen_var_9,
+        &random_rational_unsigned_unsigned_rounding_mode_quadruple_gen_var_9,
+        &special_random_rational_unsigned_unsigned_rounding_mode_quadruple_gen_var_9,
+    )
+}
+
 // All `(Rational, i64, u64, RoundingMode)` that are valid inputs to
 // `Float::root_s_rational_prec_round`.
 pub fn rational_signed_unsigned_rounding_mode_quadruple_gen_var_2()